requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.2
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
Werkzeug==2.2.2
